app/requirements/base.txt
Django==2.2
django-cors-headers==2.1.0
djangorestframework==3.9
django-filter==1.1.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
gunicorn==19.7.1
Pillow==5.1.0
celery==4.1.1
https://github.com/formatcom/django-celery-beat/zipball/v1.1.1#egg=django-celery-beat
https://github.com/formatcom/django-celery-results/zipball/v1.1.0#egg=django-celery-results
redis==2.10.6
django-rest-auth==0.9.5
cloudinary==1.11.0
kombu==4.3.0
channels==1.1.5
asgi-redis==1.4.2
asgiref==1.1.2
daphne==1.3.0
requests==2.13.0
requests-toolbelt==0.8.0
django-rest-swagger==2.2.0
static==1.1.1
static3==0.7.0
dj-static==0.0.6
sendgrid==6.1.0
openpyxl==3.0.2
pandas==1.1.5
boto3==1.12.39
botocore==1.15.39
pycryptodome==3.10.1
django-storages==1.9.1
boto3==1.12.39
botocore==1.15.39
s3transfer==0.3.3
